+++ public/sitemap-0.xml
@@ -1,11 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.core-point.kr/api/store/metadata</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/api/store/new</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/api/store/recommended</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/consulting/opening</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/consulting/transfer</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/sentry-example-page</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.core-point.kr/partnership</loc><lastmod>2023-12-12T17:55:26.006Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.core-point.kr/consulting/opening</loc><lastmod>2023-12-14T08:04:28.952Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.core-point.kr</loc><lastmod>2023-12-14T08:04:28.952Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.core-point.kr/partnership</loc><lastmod>2023-12-14T08:04:28.952Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.core-point.kr/sentry-example-page</loc><lastmod>2023-12-14T08:04:28.952Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.core-point.kr/consulting/transfer</loc><lastmod>2023-12-14T08:04:28.952Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
